update rpm build process
We would like to be able to 'make rpm' from a clean git checkout, in
order to setup an automated process for building & installing mhvtl. We
use the upstream 'git' structure for spec templating and tar/rpm make
rules. We use git-archive to grab a clean set of sources and avoid the
symlink and other hardcoded version/naming schemes.