kmod builds to use modern kmodtool
We need our kmod RPM builds for mhvtl to behave like the rest of modern
society. Our custom kmodtool wasn't a clean way to setup the builds when
we do mock builds (coming) later, as it relied on the local path and not
copying around with SOURCE.

In addition, there is no functional reason to use our own anymore. We
can use the one distributed with EPEL & redhat-rpm-config.

We dump the weak-updates config (not needed) and other file installs, as
the kmodtool rpmtemplate does %files for us and those are not included.
The files are not needed either, so no sense finding a way around
kmodtool at this time.