use mock for RPM builds
We want our builds to be host-platform agnostic, so we use our standard
mock_rpmbuild scripting. We use the VSM tricks for KVERSION setting, as
that is a bit silly but we need a kversion in the RPM that matches the
repo we are building against.

There are some notable changes, with an eye to the future for other RPM
builds. We move to using a .tar.gz as the single input, no more messing
with raw SOURCES. This greatly simplifies the variations across
different repos and puts us further toward a common mock_rpmbuild.sh