Update mock builds to handle funky kernel versions
I was hitting build failures on the latest Centos 6.x. To make a lot of
problems go away, we'll update some of the logic from the latest VSM rpm
build script. Along with this, for mhvtl we can drop back to the
standard (included with mock) epel mock configurations.

In particular, we needed to pass down KVERSION to tell mock what to look
for. This is the standard game we play, as the kversion handling in the
spec is not very sophisticated.

With this, the following command will build the complete set of RPMS for
C6 and C7:
 $ for distro in 6.x 7.x; do DISTRO_VERS=$distro make rpm kmod-rpm; done