Deploy RPMS to Artifactory
We should upload our RPMS so people can install them. We'll use the
jfrog cli as that is the easiest to automate the account details and
upload scripting.

We are using a script to deploy, as that allows us to use conditional
deploys on branches or tags. after_success would not.